Cleaning up the css files and component files for comparing players
