+++ src/components/GetFirstAndSecondPlayerStats.module.css
@@ -1,0 +1,181 @@
+.compare-players-box {
+    width: 500px;
+    border: 1px solid #c4c4c4;
+    height: 700px;
+   display: flex;
+   flex-direction: column;
+   box-shadow: 0 0 20px rgba(0, 0, 0, 0.2);
+   border-radius: 10px;
+}
+
+.compare-players-box:hover {
+    background-color: whitesmoke;
+    cursor: pointer;
+}
+
+.input-wrapper {
+    background-color: white;
+    width: 100%;
+    border-radius: 10px;
+    height: 2.5rem;
+    padding: 0 15px;
+    box-shadow: 0px 0px 8px #ddd;
+    display: flex;
+    align-items:center;
+}
+
+.input {
+    background-color: transparent;
+    border: none;
+    font-size: 1.05rem;
+    width: 100%;
+    height: 100%;
+    margin-left: 10px;
+}
+
+.input:focus{
+    outline: none;
+}
+
+.first-player-form {
+    margin: 0 auto;
+    margin-top: 50px;
+    display: flex;
+      align-items: center;
+}
+
+.first-player-search-results{
+    margin-top: 20px;
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    overflow: auto;
+}
+
+.first-player-loading-animation {
+    width: 100%;
+    transform:scale(0.5);
+}
+
+.second-player-loading-animation {
+    width: 100%;
+    transform:scale(0.5);
+}
+
+.second-player-form {
+    margin: 0 auto;
+    margin-top: 50px;
+    display: flex;
+      align-items: center;
+}
+
+.second-player-search-results{
+    margin-top: 20px;
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    overflow: auto;
+}
+
+.search-result-box {
+    border: 1px solid grey;
+    padding: 10px;
+}
+
+.search-result-box:hover {
+    background-color: whitesmoke;
+    cursor: pointer;
+}
+
+.first-player-stats-summary{
+    text-align: center;
+    margin-top: 50px;
+}
+
+.second-player-stats-summary{
+    text-align: center;
+    margin-top: 50px;
+}
+
+.compare-players-header {
+    background-color: #272727;
+    color: #fff;
+    margin-top: 50px;
+    padding: 10px;
+    text-align: center;
+    font-size: 22px;
+    font-weight: bold;
+    border-bottom: 1px solid black;
+  }
+
+.player-not-active {
+    text-align: center;
+    font-size: 20px;
+    margin-top: 50px;
+}
+
+.compare-players-image-box{
+    height: 150px;
+  
+}
+
+img {
+    width: 250px;  
+}
+
+
+
+.player-last-ten-games-stats{
+    display: flex;
+    align-items: center;
+    height: 80px;
+}
+
+
+.player-last-ten-games-stat{
+    padding: 15px;
+    border: 1px solid #c4c4c4;
+    text-align: center;
+    height: 100%;
+    width: 10%;
+}
+
+.player-last-ten-games-stat-number {
+    font-size: 20px;
+}
+
+.comparison-graph {
+    margin-top: 40px;
+    margin-bottom: 100px;
+}
+
+
+.graph-option-button {
+    margin-left: 5px;
+    border: 1px solid #7a7a7a;
+    border-radius: 10px;
+    padding: 5px 10px;
+    cursor: pointer;
+    background-color: white;
+    color: black;
+    box-shadow: 0px 0px 8px #d2d2d2;
+}
+
+.graph-option-button:hover {
+    background-color: whitesmoke;
+    color: black;
+    transform: scale(1.05);
+}
+
+.btn-selected {
+    background-color: #121212;
+    color: white;
+}
+
+.btn-selected:hover {
+    background-color: #121212;
+    color: white;
+    transform: scale(1.05); 
+}
+
+.line-graph {
+    margin-top: 30px;
+}
