+++ src/components/PastGameStats.css
@@ -22,24 +22,27 @@
     left: 50%;
     top: 50%;   
     transform: translate(-50%, 600%);
-    transition: transform 5s;
+    opacity: 1;
+    transition: opacity 5s ease-in-out;
 }
 
 /* Transition for form/input */
 .date-search-div-enter {
-    transform: translate(-50%, 600%);
+    opacity: 0;
+    transform: translateX(-50%, 600%);
 }
 
 .date-search-div-enter-active {
-    transform: translate(-50%, -850%);
-    height: 50px;
-    transition: transform 500ms ease-in-out;
+    transform: translateY(-50%, -800%);
+    /* height: 50px; */
+    opacity: 0;
+    transition: opacity 500ms ease-in-out;
 }
 
 .date-search-div-enter-done {
     position: absolute;
-        height: 50px;
-    transform: translate(-50%, -850%);
+    transform: translate(-50%, -800%);
+    opacity: 1;
 }
 
 h1 {
@@ -52,7 +55,6 @@ h1 {
     grid-template-columns: repeat(2, 1fr); 
     gap: 1rem; 
     margin-bottom: 100px;
-
 }
 
 
@@ -94,7 +96,7 @@ h1 {
     margin-top: 100px;
     margin-bottom: 100px;
     height: 500px;
-    overflow: auto;
+    /* overflow: auto; */
 }
 
 .fade-in {
